google/link, intel/wtm2: Do not include non-existent "hda.h" in i915.c
Commit 75c83870 (azalia: Shrink boilerplate) [1] removed the header
file `hda_verb.h`. This header is still included in the mainboard’s
`i915.c`, causing the following build error, when native graphics
initialization is enabled.

        CC         mainboard/intel/wtm2/i915.ramstage.o
    src/mainboard/intel/wtm2/i915.c:34:22: fatal error: hda_verb.h: No such file or directory

This was not caught, as native graphics initialization is not enabled
for the build tests.

It turns out that the array `mainboard_cim_verb_data` is not used in
`src/mainboard/intel/wtm2/hda_verb.h`, so fix the problem by removing
the inclusion.

[1] http://review.coreboot.org/6840

Change-Id: Ic902581c6809a1069e169cc874678146a24d75f3
Signed-off-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-on: http://review.coreboot.org/6945
Tested-by: build bot (Jenkins)
Reviewed-by: Vladimir Serbinenko <phcoder@gmail.com>
Reviewed-by: Aaron Durbin <adurbin@google.com>