Doxygen: add a "simple" output config and make target
This is a doxygen target I'm using for cleaning up the coreboot doxygen
output.  It runs in about a minute instead of....  a lot longer, and
only generates documentation for the coreboot/src directory, excluding
vendorcode.

It was requested that the comments be stripped to make it easier to
read.  This was done through the following command (split for gerrit):

head -n 1 Doxyfile.coreboot_simple > Doxyfile.tmp ;
cat Doxyfile.coreboot_simple | sed 's|^\s*#.*$||' | sed '/^$/d' | sed 's|[[:space:]]\+$||' >> Doxyfile.tmp ;
mv Doxyfile.tmp Doxyfile.coreboot_simple

I'm including the command here because any time the file is updated
with the doxygen wizard, it will need to be run again.  It might be
desirable to turn this into a script in the documentation directory.

Change-Id: Ic0cbbcd21aa647e80a037481546bdcd2aab8949e
Signed-off-by: Martin Roth <gaumless@gmail.com>
Reviewed-on: http://review.coreboot.org/8068
Tested-by: build bot (Jenkins)
Reviewed-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>